+++ jans-core/cedarling/pom.xml
@@ -90,10 +90,6 @@
 		</dependency>
 		<dependency>
 			<groupId>io.jans</groupId>
-			<artifactId>jans-auth-client</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>io.jans</groupId>
 			<artifactId>cedarling-java</artifactId>
 		</dependency>
