+++ jans-lock/lock-server/service/src/main/java/io/jans/lock/service/ws/rs/stat/StatRestWebService.java
@@ -1,6 +1,6 @@
 package io.jans.lock.service.ws.rs.stat;
 
-import io.jans.lock.cedarling.service.security.api.ProtectedCedarlingApi;
+import io.jans.core.cedarling.service.security.api.ProtectedCedarlingApi;
 import io.jans.lock.model.core.LockApiError;
 import io.jans.lock.model.stat.FlatStatResponse;
 import io.jans.lock.util.ApiAccessConstants;
